Move IngameMenuHandler methods into SaveUtils where they belong.
